Move MiniTestHeckler to its own file.
